#415: P2R - fix AuditlogEntryDTO export
